extract docmap fetchin'
