+++ packages/widget/src/index.ts
@@ -1,2 +1,3 @@
 // All files which will be accessible when the widget is installed via npm are declared here
 export * from './docmaps-widget'
+export { getDocmap } from './docmap-controller.ts';
